Remove usages of deprecated 'window-system' variable
As per docstring, the variable has been deprecated.  Where it makes
sense, I have replaced it with 'display-graphic-p'.  When declaring
`agda2-fontset-name` the variable was compared to the symbol `mac`,
which per docstring and (elisp) Window Systems was never a legal value
(perhaps `system-type` was meant?), so I have simplified the default
value, as this has apparently not been affecting anyone since 2009.